some restructuring of the code and changes to power to agree better with nbodykit (some numerical instability; ask Lehman); attempt at adding classy and ZeNBu (unclear how to add a pip install from git as an optional requirement); added tests for power spectrum and for zcv
