+++ abacusnbody/hod/abacus_hod.py
@@ -257,7 +257,8 @@ from astropy.io import ascii
 
 from .GRAND_HOD import *
 from .parallel_numpy_rng import *
-from .tpcf_corrfunc import calc_xirppi_fast, calc_wp_fast, calc_multipole_fast, calc_power
+from .tpcf_corrfunc import calc_xirppi_fast, calc_wp_fast, calc_multipole_fast
+from .power_spectrum import calc_power
 from .zcv.advect_fields import advect
 from .zcv.tools_jdr import run_zcv
 # TODO B.H.: staging can be shorter and prettier; perhaps asdf for h5 and ecsv?
@@ -1019,12 +1020,15 @@ class AbacusHOD:
         clustering: dict
             dictionary of summary statistics. Auto-correlations/spectra can be
             accessed with keys such as ``'LRG_LRG'`` and ``'LRG_LRG_ell'`` for the 
-            multipoles. Cross-correlations/spectra can be accessed with keys such 
-            as ``'LRG_ELG'`` and ``'LRG_ELG_ell'`` for the multipoels. Keys ``k_binc`` 
+            multipoles with number of modes per bin, ``'LRG_LRG[_ell]_modes'``. 
+            Cross-correlations/spectra can be accessed with keys such 
+            as ``'LRG_ELG'`` and ``'LRG_ELG_ell'`` for the multipoles 
+            with number of modes per bin, ``'LRG_LRG[_ell]_modes'``. Keys ``k_binc`` 
             and ``mu_binc`` contain the bin centers of k and mu, respectively. 
             The power spectrum P(k, mu) has a shape (nbins_k, nbins_mu), whereas
-            the multipole power spectrum has shape (len(poles), nbins_k).
+            the multipole power spectrum has shape (len(poles), nbins_k). Cubic box only.
         """
+        Lbox = self.lbox
         clustering = {}
         for i1, tr1 in enumerate(mock_dict.keys()):
             x1 = mock_dict[tr1]['x']
@@ -1035,54 +1039,77 @@ class AbacusHOD:
                 if i1 > i2: continue # cross-correlations are symmetric
                 if i1 == i2:
                     print(tr1+'_'+tr2)
-                    k_binc, mu_binc, pk3d, N3d, binned_poles, Npoles = calc_power(x1, y1, z1, nbins_k, nbins_mu, k_hMpc_max, logk, self.lbox, paste, num_cells, compensated, interlaced, w = w1, poles = poles)
+                    k_binc, mu_binc, pk3d, N3d, binned_poles, Npoles = calc_power(x1, y1, z1, nbins_k, nbins_mu, k_hMpc_max, logk, Lbox, paste, num_cells, compensated, interlaced, w = w1, poles = poles)
                     clustering[tr1+'_'+tr2] = pk3d
+                    clustering[tr1+'_'+tr2+'_modes'] = N3d
                     clustering[tr1+'_'+tr2+'_ell'] = binned_poles
+                    clustering[tr1+'_'+tr2+'_ell_modes'] = Npoles
                 else:
                     print(tr1+'_'+tr2)
                     x2 = mock_dict[tr2]['x']
                     y2 = mock_dict[tr2]['y']
                     z2 = mock_dict[tr2]['z']
                     w2 = mock_dict[tr2].get('w', None)
-                    k_binc, mu_binc, pk3d, N3d, binned_poles, Npoles = calc_power(x1, y1, z1, nbins_k, nbins_mu, k_hMpc_max, logk, self.lbox, paste, num_cells, compensated, interlaced,
+                    k_binc, mu_binc, pk3d, N3d, binned_poles, Npoles = calc_power(x1, y1, z1, nbins_k, nbins_mu, k_hMpc_max, logk, Lbox, paste, num_cells, compensated, interlaced,
                                                       w = w1, x2 = x2, y2 = y2, z2 = z2, w2 = w2, poles = poles)
                     clustering[tr1+'_'+tr2] = pk3d
+                    clustering[tr1+'_'+tr2+'_modes'] = N3d
                     clustering[tr1+'_'+tr2+'_ell'] = binned_poles
+                    clustering[tr1+'_'+tr2+'_ell_modes'] = Npoles
                     clustering[tr2+'_'+tr1] = clustering[tr1+'_'+tr2]
+                    clustering[tr2+'_'+tr1+'_modes'] = clustering[tr1+'_'+tr2+'_modes']
                     clustering[tr2+'_'+tr1+'_ell'] = clustering[tr1+'_'+tr2+'_ell']
+                    clustering[tr2+'_'+tr1+'_ell_modes'] = clustering[tr1+'_'+tr2+'_ell_modes']
         clustering['k_binc'] = k_binc
         clustering['mu_binc'] = mu_binc
         return clustering
 
-    def apply_zcv(self, mock_dict, config):
+    def apply_zcv(self, mock_dict, config, load_presaved=False):
         """
         Apply control variates reduction of the variance to a power spectrum observable.
         """
         # compute real space and redshift space
-        assert config['HOD_params']['want_rsd'], "Currently want_rsd=False not implemented"
-        assert len(mock_dict.keys()) == 1, "Currently implemented only a single tracer"
-        assert len(config['power_params']['poles']) == 3, "Currently implemented only multipoles 0, 2, 4"
-        
-        # run version with rsd (make pretty)
-        for tr in mock_dict.keys():
-            # obtain the positions
-            tracer_pos = np.vstack((mock_dict[tr]['x'], mock_dict[tr]['y'], mock_dict[tr]['z'])).T
-            del mock_dict; gc.collect()
+        #assert config['HOD_params']['want_rsd'], "Currently want_rsd=False not implemented"
+        assert len(mock_dict.keys()) == 1, "Currently implemented only a single tracer" # should make a dict of dicts, but need cross
+        assert len(config['power_params']['poles']) == 3, "Currently implemented only multipoles 0, 2, 4; need to change ZeNBu"
 
-            # advect fields for this tracer
-            pk_rsd_tr_dict, pk_rsd_ij_dict = advect(tracer_pos, config['HOD_params']['want_rsd'], config)
-        
-        # run version without rsd
-        mock_dict = self.run_hod(self.tracers, want_rsd=False, reseed=None, write_to_disk=False, 
-                                 Nthread=16, verbose=False, fn_ext=None)
-        for tr in mock_dict.keys():
-            # obtain the positions
-            tracer_pos = np.vstack((mock_dict[tr]['x'], mock_dict[tr]['y'], mock_dict[tr]['z'])).T
-            del mock_dict; gc.collect()
+        if load_presaved:
+            # create save directory
+            save_dir = Path(config['zcv_params']['zcv_dir']) / config['sim_params']['sim_name']
+            save_z_dir = save_dir / f"z{config['sim_params']['z_mock']:.3f}"
+            rsd_str = "_rsd" if config['HOD_params']['want_rsd'] else ""
+            pk_rsd_tr_dict = asdf.open(save_z_dir / f"power{rsd_str}_tr_nmesh{config['zcv_params']['nmesh']}.asdf")['data']
+            pk_rsd_ij_dict = asdf.open(save_z_dir / f"power{rsd_str}_ij_nmesh{config['zcv_params']['nmesh']}.asdf")['data']
+            if config['HOD_params']['want_rsd']:
+                pk_tr_dict = asdf.open(save_z_dir / f"power_tr_nmesh{config['zcv_params']['nmesh']}.asdf")['data']
+                pk_ij_dict = asdf.open(save_z_dir / f"power_ij_nmesh{config['zcv_params']['nmesh']}.asdf")['data']
+            else:
+                pk_tr_dict, pk_ij_dict = None, None
+                        
+        else:
+            # run version with rsd or without rsd
+            for tr in mock_dict.keys():
+                # obtain the positions
+                tracer_pos = np.vstack((mock_dict[tr]['x'], mock_dict[tr]['y'], mock_dict[tr]['z'])).T
+                del mock_dict; gc.collect()
 
-            # advect fields for this tracer
-            pk_tr_dict, pk_ij_dict = advect(tracer_pos, want_rsd=False, config=config)
+                # advect fields for this tracer
+                pk_rsd_tr_dict, pk_rsd_ij_dict = advect(tracer_pos, config['HOD_params']['want_rsd'], config)
 
+            # run version without rsd if rsd was requested
+            if config['HOD_params']['want_rsd']:
+                mock_dict = self.run_hod(self.tracers, want_rsd=False, reseed=None, write_to_disk=False, 
+                                         Nthread=16, verbose=False, fn_ext=None)
+                for tr in mock_dict.keys():
+                    # obtain the positions
+                    tracer_pos = np.vstack((mock_dict[tr]['x'], mock_dict[tr]['y'], mock_dict[tr]['z'])).T
+                    del mock_dict; gc.collect()
+
+                    # advect fields for this tracer
+                    pk_tr_dict, pk_ij_dict = advect(tracer_pos, want_rsd=False, config=config)
+            else:
+                pk_tr_dict, pk_ij_dict = None, None
+                
         # run the final part and save
         zcv_dict = run_zcv(pk_rsd_tr_dict, pk_rsd_ij_dict, pk_tr_dict, pk_ij_dict, config)
         return zcv_dict
